backend/requirements.txt
asgiref==3.10.0
black==25.9.0
cfgv==3.4.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.0
flake8==7.3.0
fonttools==4.60.1
graphviz==0.21
gunicorn==23.0.0
identify==2.6.15
iniconfig==2.3.0
isort==7.0.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.35.3
